% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/arpa.R
\docType{package}
\name{arpa}
\alias{arpa}
\alias{arpa-package}
\title{arpa: A package for parsing and manipulating ngram models in ARPA format}
\description{
The arpa package includes the S4 class \linkS4class{ngram.model}
}
\details{
Learn more at the project homepage:
\url{http://github.com/adamacosta/arpa/wiki}
}
\author{
Adam Acosta, Sam Acosta
}

