% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram.model.methods.R
\docType{methods}
\name{bigrams,ngram.model-method}
\alias{bigrams,ngram.model-method}
\title{Return the bigrams of a language model}
\usage{
\S4method{bigrams}{ngram.model}(object)
}
\arguments{
\item{object}{An ngram.model object}
}
\value{
bigrams The bigrams of the model
}
\description{
Return the bigrams of a language model
}
\author{
Adam Acosta
}

