% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram.model.methods.R
\name{bigrams}
\alias{bigrams}
\title{Return the bigrams of a language model}
\usage{
bigrams(object)
}
\arguments{
\item{object}{An ngram.model object}
}
\value{
bigrams The bigrams of the model
}
\description{
Return the bigrams of a language model
}
\author{
Adam Acosta
}

