% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tokenize.R
\name{clean_text}
\alias{clean_text}
\title{Return cleaned version of input text}
\usage{
clean_text(text)
}
\arguments{
\item{text}{Unformatted character vector}
}
\value{
tokenizable character vectors
}
\description{
Return cleaned version of input text
}
\author{
Adam Acosta
}

