% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram.model.methods.R
\name{contains}
\alias{contains}
\title{Tests whether or not a language model contains an ngram}
\usage{
contains(object, key)
}
\arguments{
\item{object}{An ngram.model object}

\item{key}{A character string}
}
\value{
boolean Whether or not the string is in the language model
}
\description{
Tests whether or not a language model contains an ngram
}
\author{
Adam Acosta
}

