% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram.model.R
\docType{class}
\name{ngram.model-class}
\alias{ngram.model}
\alias{ngram.model-class}
\title{An S4 class to represent an ngram language model}
\description{
Provides an efficient, fast way to map ngrams in a language
model to their log probability, allowing for easy next-word prediction.
}
\section{Slots}{

\describe{
\item{\code{unigrams}}{A data.table mapping unigrams to their log probability}

\item{\code{bigrams}}{A data.table mapping bigrams to their log probability}

\item{\code{trigrams}}{A data.table mapping trigrams to their log probability}
}}
\author{
Adam Acosta
}

