% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{parseFileR}
\alias{parseFileR}
\title{Helper called by read.arpa}
\usage{
parseFileR(input, uskip, umax, bskip, bmax, tskip, tmax, qskip, qmax,
  small = TRUE)
}
\arguments{
\item{input}{A filename}

\item{uskip}{numeric}

\item{umax}{numeric}

\item{bskip}{numeric}

\item{bmax}{numeric}

\item{tskip}{numeric}

\item{tmax}{numeric}
}
\description{
Helper called by read.arpa
}

