% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict.next,ngram.model,character-method}
\alias{predict.next,ngram.model,character-method}
\title{Return the predicted next word in a sentence}
\usage{
\S4method{predict.next}{ngram.model,character}(object, text)
}
\arguments{
\item{object}{ngram.model An ngram.model object}

\item{text}{character A string of input text}
}
\value{
word character The predicted next word
}
\description{
Return the predicted next word in a sentence
}
\author{
Adam Acosta
}

