% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.next.R
\name{predict.next}
\alias{predict.next}
\title{Word prediction engine}
\usage{
predict.next(object, text)
}
\arguments{
\item{object}{ngram.model}

\item{text}{character}
}
\value{
The most likely next word in the sentence
}
\description{
Receives input text and predicts what the next word will be.
}
\author{
Adam Acosta
}

