% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,ngram.model,character-method}
\title{Word prediction engine}
\usage{
predict(object, text)

\S4method{predict}{ngram.model,character}(object, text)
}
\arguments{
\item{object}{ngram.model}

\item{text}{character}
}
\value{
The most likely next word in the sentence
}
\description{
Receives input text and predicts what the next word will be.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = ngram.model,text = character}: Next-word prediction engine
}}
\author{
Adam Acosta

Adam Acosta
}

