% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram.model.methods.R
\name{prob}
\alias{prob}
\title{Returns the log probability of an ngram in a language model}
\usage{
prob(object, key)
}
\arguments{
\item{object}{An ngram.model object}

\item{key}{A character string}
}
\value{
numeric The log probability of the provided ngram
}
\description{
Returns the log probability of an ngram in a language model
}
\author{
Adam Acosta
}

