% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.arpa.R
\name{read.arpa}
\alias{read.arpa}
\title{Read ARPA file}
\usage{
read.arpa(input = "", header = TRUE, verbose = FALSE, nrow = -1L,
  skip = 0L, ugrams = -1L, bgrams = -1L, tgrams = -1L)
}
\arguments{
\item{input}{A filename}

\item{header}{boolean indicating whether or not the file has a header.
Default is TRUE}

\item{verbose}{A boolean indicating whether you want the function to
print information to the console as it is parsing the file. Default is FALSE.}

\item{nrow}{The number of rows in the file. This is optional, as the format
itself dictates that the file header must give the number of each ngram,
from which the number of lines in the file can be inferred. If this is passed
as a parameter, pass the number of unigrams, bigrams, and trigrams as well,
and it will speed up the parsing process.}

\item{skip}{The number of rows, if any, to skip in the file. Default is 0.}

\item{ugrams}{See below}

\item{bgrams}{See below}

\item{tgrams}{integer values indicating the number of unigrams, bigrams,
and trigrams. Should be in the header but can be passed to the function
directly to avoid header parsing overhead.}
}
\value{
An ngram.model object, stored internally as a list of three hash
tables mapping each ngram to its log probability.
}
\description{
Reads an ARPA file and returns the language model. See
http://www.speech.sri.com/projects/srilm/manpages/ngram-format.5.html
for a description of the ARPA file format.
}
\author{
Adam Acosta
}

