% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram.model.methods.R
\name{subset}
\alias{subset}
\title{Return the subset of the language model contain the partial key}
\usage{
subset(object, key)
}
\arguments{
\item{object}{An ngram.model object}

\item{key}{A character string}
}
\value{
data.table The subset of the ngram.model containing key
}
\description{
Return the subset of the language model contain the partial key
}
\author{
Adam Acosta
}

