% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram.model.methods.R
\name{tetragrams}
\alias{tetragrams}
\title{Return the tetragrams of a language model}
\usage{
tetragrams(object)
}
\arguments{
\item{object}{An ngram.model object}
}
\value{
tetragrams The tetragrams of the model
}
\description{
Return the tetragrams of a language model
}
\author{
Adam Acosta
}

