% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\title{Return vector of tokens from input text}
\usage{
tokenize(text)
}
\arguments{
\item{text}{Unformatted character vector}
}
\value{
words A list of tokens parsed from the input
}
\description{
Return vector of tokens from input text
}
\author{
Adam Acosta
}

