% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram.model.methods.R
\name{trigrams}
\alias{trigrams}
\title{Return the trigrams of a language model}
\usage{
trigrams(object)
}
\arguments{
\item{object}{An ngram.model object}
}
\value{
trigrams The trigrams of the model
}
\description{
Return the trigrams of a language model
}
\author{
Adam Acosta
}

