% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ngram.model.methods.R
\name{unigrams}
\alias{unigrams}
\title{Return the unigrams of a language model}
\usage{
unigrams(object)
}
\arguments{
\item{object}{An ngram.model object}
}
\value{
unigrams The unigrams of the model
}
\description{
Return the unigrams of a language model
}
\author{
Adam Acosta
}

