% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-ngram.model.R, R/ngram.model.methods.R
\docType{class}
\name{ngram.model-class}
\alias{bigrams,ngram.model-method}
\alias{contains,ngram.model,character-method}
\alias{ngram.model}
\alias{ngram.model-class}
\alias{prob,ngram.model,character-method}
\alias{subset,ngram.model,character-method}
\alias{tetragrams,ngram.model-method}
\alias{trigrams,ngram.model-method}
\alias{unigrams,ngram.model-method}
\title{An S4 class to represent an ngram language model}
\usage{
\S4method{unigrams}{ngram.model}(object)

\S4method{bigrams}{ngram.model}(object)

\S4method{trigrams}{ngram.model}(object)

\S4method{tetragrams}{ngram.model}(object)

\S4method{contains}{ngram.model,character}(object, key)

\S4method{subset}{ngram.model,character}(object, key)

\S4method{prob}{ngram.model,character}(object, key)
}
\arguments{
\item{object}{An ngram.model object}

\item{key}{A character string}
}
\description{
Provides an efficient, fast way to map ngrams in a language
model to their log probability, allowing for easy next-word prediction.
}
\section{Methods (by generic)}{
\itemize{
\item \code{unigrams}: Return the unigrams of a ngram.model

\item \code{bigrams}: Return the bigrams of a ngram.model

\item \code{trigrams}: Return the trigram of a ngram.model

\item \code{tetragrams}: Return the tetragrams of a ngram.model

\item \code{contains}: Test whether or not a ngram.model contain an ngram

\item \code{subset}: Return the subset of an ngram.model containing a partial key

\item \code{prob}: Return the log probability of an ngram in a ngram.model
}}
\section{Slots}{

\describe{
\item{\code{unigrams}}{A data.table mapping unigrams to their log probability}

\item{\code{bigrams}}{A data.table mapping bigrams to their log probability}

\item{\code{trigrams}}{A data.table mapping trigrams to their log probability}
}}
\author{
Adam Acosta

Adam Acosta

Adam Acosta

Adam Acosta

Adam Acosta

Adam Acosta

Adam Acosta

Adam Acosta
}

