% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.next.R
\docType{methods}
\name{predict.next,ngram.model,character-method}
\alias{predict.next,ngram.model,character-method}
\title{inheritParams predict.next}
\usage{
\S4method{predict.next}{ngram.model,character}(object, text)
}
\description{
inheritParams predict.next
}

