% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.next.R
\docType{methods}
\name{predict.next,ngram.model,character-method}
\alias{predict.next,ngram.model,character-method}
\title{Word prediction engine}
\usage{
\S4method{predict.next}{ngram.model,character}(object, text)
}
\arguments{
\item{object}{ngram.model}

\item{text}{character}
}
\value{
next character
}
\description{
Receives input text and predicts what the next word will be.
}
\author{
Adam Acosta
}

